Old outliner removed
