+++ src/interfaces/OutlinerTypes.ts
@@ -1,0 +1,7 @@
+export type OutlinerTypes =
+  | "project"
+  | "package"
+  | "materialgraph"
+  | "datagraph"
+  | "shadergraph"
+  | "kernelgraph";
