--- src/renderer/scss/outliner.scss
@@ -1,11 +1,0 @@
-@import "./colors.scss";
-
-.treeItem {
-  cursor: pointer;
-}
-
-.treeItemActive {
-  cursor: pointer;
-  background-color: $FONT_COLOR;
-  color: $IMPORTANT_BACKGROUND_COLOR;
-}
